Extract digestAuthenticator
