+++ core/shared/src/test/scala/sttp/client/DigestAuthenticatorTest.scala
@@ -1,0 +1,7 @@
+package sttp.client
+
+import org.scalatest.{FlatSpec, Matchers}
+
+class DigestAuthenticatorTest extends FlatSpec with Matchers {
+  it should "asd" in {}
+}
